backend/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
python-dotenv==1.0.1
psycopg[binary]==3.2.4